backend

+++ backend/controllers/goalControllers.js
@@ -1,8 +1,9 @@
 const asyncHandler = require('express-async-handler') 
 const Goal = require('../models/goalsModel')
-
+const User = require('../models/usersModel')
+ 
 const getGolas = asyncHandler (async (reg, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({user: reg.user.id})
     res.json({goals})
 })
 
@@ -14,6 +15,7 @@ const setGoal = asyncHandler (async(reg, res) => {
 
     const goal = await Goal.create({
         text: reg.body.text,
+        user: reg.user.id
     })
 
     res.json({goal})
@@ -26,6 +28,18 @@ const updateGoal = asyncHandler (async(reg, res) => {
     if(!goal){
         res.status(400)
         throw new Error('Goal not updated')
+    }
+
+    const user = await User.findById(reg.user.id)
+
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+
+    if(!user){
+        res.status(401)
+        throw new Error('User Not Found')
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(reg.params.id, reg.body, {
@@ -41,7 +55,19 @@ const deleteGoal = asyncHandler (async (reg, res) => {
 
     if(!goal){
         res.status(400)
-        throw new Error('Goa to remove not found')
+        throw new Error('Goal to remove not found')
+    }
+
+    const user = await User.findById(reg.user.id)
+
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+
+    if(!user){
+        res.status(401)
+        throw new Error('User Not Found')
     }
 
     await goal.deleteOne()
